update live data location including current position
